minor changes so tests pass
